api: sched: split defaults from normalize
no user-visible changes. Internal code movement only.
